shinychat: Chat UI Component for 'shiny'

Provides a scrolling chat interface with multiline input, suitable for creating chatbot apps based on Large Language Models (LLMs). Designed to work particularly well with the 'ellmer' R package for calling LLMs.

See also

Author

Maintainer: Garrick Aden-Buie garrick@adenbuie.com (ORCID)

Authors:

Joe Cheng joe@posit.co

Carson Sievert carson@posit.co

Barret Schloerke barret@posit.co (ORCID)

Other contributors:

Posit Software, PBC (ROR) [copyright holder, funder]