shinychat: Chat UI Component for 'shiny'

Provides a scrolling chat interface with multiline input, suitable for creating chatbot apps based on Large Language Models (LLMs). Designed to work particularly well with the 'ellmer' R package for calling LLMs.

See also

Author

Maintainer: Joe Cheng joe@posit.co

Authors:

Carson Sievert carson@posit.co

Other contributors:

Posit Software, PBC [copyright holder, funder]